# .stryker-tmp/sandbox-bCteHS/docs/api-contract.yaml
openapi: 3.0.3
info:
  title: VibeX API Contract
  description: |
    VibeX 全栈项目 API 契约文档
    
    ## 版本历史
    - v1.1.0 (2026-03-28): 从 14 端点扩展至 90+ 端点，覆盖全部前端实际调用
    - v1.0.0 (初始): 基础 14 端点定义
    
    ## 路径约定
    - 所有路径相对于 `/api` 前缀（如 `/auth/login` → `/api/auth/login`）
    - 服务器基础路径: `/api`
    - 统一无 v1 前缀（v1 双写路由已废弃，标注 `deprecated: true`）
    
    ## 路由分类
    - **前端调用**: 前端 services/api/ 模块中实际调用
    - **后端独有**: `x-backend-only: true`，前端未调用但后端已实现
    - **废弃路由**: `deprecated: true`，v1 双写兼容层
    
    ## 16 个 Tag 分组
    | Tag | 前缀 | 说明 |
    |-----|------|------|
    | Auth | /auth | 认证 |
    | Users | /users | 用户 |
    | Projects | /projects | 项目管理 |
    | Requirements | /requirements | 需求 |
    | Clarifications | /clarifications | 澄清问答 |
    | DomainEntities | /domain-entities, /domains | 域实体 |
    | EntityRelations | /entity-relations | 实体关系 |
    | DDD | /ddd | DDD 限界上下文/域模型/业务流程 |
    | Design | /clarify, /domain, /flow, /pages, /prototype, /design | 设计流程 |
    | Flows | /flows | 流程图 |
    | PrototypeSnapshots | /prototype-snapshots | 原型快照 |
    | Agents | /agents | Agent |
    | Pages | /pages | 页面 |
    | Chat | /chat | AI 对话 |
    | Plan | /plan | 计划分析 |
    | Diagnosis | /diagnosis | 需求诊断 |
    | BackendOnly | 各类后端独有路由 | 未被前端调用的后端路由 |
    | Deprecated | /api/v1/* | v1 兼容层（废弃） |
  version: 1.1.0
  contact:
    name: VibeX Team
servers:
  - url: /api
    description: API 基础路径
...
paths:
...
  /requirements/{requirementId}/domains:
    post:
      tags: [Requirements]
      summary: 为需求创建关联的域实体
      operationId: createRequirementDomain
      parameters:
        - name: requirementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainEntityCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: '#/components/schemas/DomainEntity'
        '400':
          $ref: '#/components/responses/BadRequest'

  # =============================================================================
  # TAG: CLARIFICATIONS
  # =============================================================================
...
  /domains/{entityId}:
    get:
      tags: [DomainEntities]
      summary: 获取单个域实体
      operationId: getDomainEntity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: '#/components/schemas/DomainEntity'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [DomainEntities]
      summary: 更新域实体
      operationId: updateDomainEntity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainEntityUpdate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  domain:
                    $ref: '#/components/schemas/DomainEntity'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [DomainEntities]
      summary: 删除域实体
      operationId: deleteDomainEntity
      parameters:
        - name: entityId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # =============================================================================
  # TAG: ENTITY RELATIONS
  # =============================================================================
...
  /ddd/domain-model:
    post:
      tags: [DDD]
      summary: 从限界上下文生成域模型
      operationId: generateDomainModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [boundedContexts, requirementText]
              properties:
                boundedContexts:
                  type: array
                  items:
                    $ref: '#/components/schemas/BoundedContext'
                requirementText:
                  type: string
                projectId:
                  type: string
      responses:
        '200':
          description: 生成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainModelResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /domain/generate:
    post:
      tags: [Design]
      summary: 生成域模型（设计流程）
      operationId: generateDomainModelDesign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DomainModelRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainModelResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /domain-models:
    post:
      tags: [BackendOnly]
      summary: 创建域模型记录
      operationId: createDomainModel
      x-backend-only: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                projectId:
                  type: string
                name:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  domainModel:
                    type: object
    get:
      tags: [BackendOnly]
      summary: 获取域模型列表
      operationId: getDomainModels
      x-backend-only: true
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  domainModels:
                    type: array
                    items:
                      type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DomainEntity:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [aggregate, entity, value-object, domain-event]
        description:
          type: string
          nullable: true
        requirementId:
          type: string
          format: uuid
          nullable: true
        attributes:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              type:
                type: string
              required:
                type: boolean
        relationships:
          type: array
          items:
            type: object
            properties:
              target:
                type: string
              type:
                type: string
              description:
                type: string
        createdAt:
          type: string
          format: date-time
    DomainEntityCreate:
      type: object
      required:
        - name
        - type
        - requirementId
      properties:
        name:
          type: string
        type:
          type: string
          enum: [aggregate, entity, value-object, domain-event]
        requirementId:
          type: string
        description:
          type: string
        attributes:
          type: array
          items:
            type: object
    DomainEntityUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [aggregate, entity, value-object, domain-event]
        description:
          type: string
        attributes:
          type: array
          items:
            type: object

    # ===== 实体关系相关 =====
...
    BoundedContext:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [core, supporting, generic, external]
        keyResponsibilities:
          type: array
          items:
            type: string
        relationships:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              fromContextId:
                type: string
              toContextId:
                type: string
              type:
                type: string
                enum: [upstream, downstream, symmetric]
              description:
                type: string
...
    DomainModelResponse:
      type: object
      properties:
        success:
          type: boolean
        domainModels:
          type: array
          items:
            type: object
        mermaidCode:
          type: string
        error:
          type: string
...
    DomainModelRequest:
      type: object
      required:
        - sessionId
        - boundedContexts
        - requirementText
      properties:
        sessionId:
          type: string
        boundedContexts:
          type: array
          items:
            type: object
        requirementText:
          type: string
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务器内部错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'